flutter_deneme/backend/requirements.txt
flask
python-dotenv
google-genai
zemberek-python
JPype1
setuptools
